poni: add get_override_config_path() to core poni PlugIn class
If you use "poni update-config" to update a file in a config, "poni
deploy" will use the new version only if a top parent config was
updated, i.e. "update-config" doesn't work on inherited nodes when
"deploy" is involved.

With this change you can use "get_override_config_path" function or
"auto_override" parameter to add_file() to mark the file as
overrideable in inherited nodes.