poni: keyword to get script path and more information in file not found situation
Added a way to expand the directory from which a script was ran inside itself.
This is now possible with "$script_path" variable which will be rendered to a value
which is full path to the script with script filename removed.

Also added more verbosity when file is not found in update-config