core: added tagging support for files and dirs
Files and directories can be optionally labelled with tags
that can be used to perform verify/audit/deploy/show operations
over a subset of the deployable files.

If a file/dir has been tagged, it is not included by default to
verify/audit/deploy/show operations. Instead, it must be explicitly
targetted via the --tag=TAG switch.

Tags can be added with a new plugin add_file() method "tags" parameter:

  self.add_file(source_file, tags=["foo", "bar", ""])

An empty string in the tag list indicates that the file shall be
included in the operation also when no --tag=TAG argument has been
specified.